backend/requirements/base.txt
Django
djangorestframework
djangorestframework-simplejwt
drf-spectacular
django-filter
django-cors-headers
mysqlclient
python-dotenv
gunicorn
requests
